Add the ability to add keywords and comments via the .ini file.
Summary:
This makes it possible to specify values to -k and -c in the .ini
file, without needing to do it on the commandline.  Further, it allows
for -k and -c options that are per extractor, rather than global.  (So
we can add '-k _js' but just for .html files.)

I had to move parse_keywords() from frontend.py to extract.py, which
is a little sad, but otherwise the change is very localized.

Test Plan:
Added to babel.ini:
   -c = I18n:    # to all extractors
   -k = _js      # to the jinja2 extractor only
and ran
   tools/pybabel extract -F tools/babel.ini .
and verified everything showed up like it should:
1) .POT comment:
  #. This means "The history of publishes": what publishes were done
  #: templates/admin-publish-status.html:60
  msgid "Publish History"
  msgstr ""
2) .js files ($._ handling):
  #: javascript/shared-package/goals.js:704
  msgid "Page load failed. Please try again."
  msgstr ""
3) _js:
  #: templates/api-tos.html:202
  msgid "Back to Top"
  msgstr ""

Reviewers: jeresig

Reviewed By: jeresig

Differential Revision: http://phabricator.khanacademy.org/D1050

Conflicts:
	babel/messages/extract.py
	babel/messages/frontend.py