Remove the requirement that --no-default-keywords be paired with -k.
Summary:
babel requires *some* keyword to look for, but now that -k can be
specified on a per-handler basis in the .ini file, it's a lot more
work for the driver script to tell whether there's always a keyword
when one is needed, or not.

So we assume the best, that if the user specifies
--no-default-keywords but doesn't specify -k, then they've got '-k'
lines up the wazoo in their .ini file.  It's better than dying (the
current behavior) which makes us jump through hoops with 'fake' global
-k flags, and it's a lot less work than actually parsing the config
file and trying to determine if -k is specified everywhere.

Maniphest: T343

Test Plan:
None.  Will run
   tools/pybabel -F tools/babel.ini --no-default-keywords .
once this is integrated into stable, to make sure it doesn't die.

Reviewers: john

Reviewed By: john

Differential Revision: http://phabricator.khanacademy.org/D2026