Speed up extraction by skipping entire directories that are marked ignore.
Summary:
If our babel.ini has a rule like this:
   [ignore: clienttemplates/**]
pybabel was iterating through every file in clienttemplates/, and then
dutifully ignoring each one.  I now notice ignore patterns of the form
'directory/**' and just remove them from the os.walk, so we skip them
entirely.

Test Plan:
Added a logging message when deleting a directory from the walk, and
then ran
  make i18n
and saw the following messages:
  ignoring directory third_party
  ignoring directory stylesheets
  ignoring directory images
  ignoring directory genfiles
  ignoring directory flvplayer
  ignoring directory compiled_templates
  ignoring directory clienttemplates
  ignoring directory node_modules
  ignoring directory images
  ignoring directory images
  ignoring directory test
  ignoring directory test
  ignoring directory images
  ignoring directory images
  ignoring directory test
  ignoring directory test
  ignoring directory node_modules
  ignoring directory images

These all look correct.  In particular, it does not ignore
translations (which is good, because we need to process
translations/manual.json), and it does not ignore khan_exercises
(which is a directory rule, but not an ignore: rule).

Reviewers: john

Reviewed By: john

Differential Revision: http://phabricator.khanacademy.org/D2087