Correctly handle the case this package might be inside a zipfile.
It does an open(datafile), which fails when we're in a zipfile,
because then our path is 'foo/bar.zip/datafile', and open() on that
fails.  (I blame open().  If we're going to have zipimport...  But
anyway.)

Now we detect if we're in a zipfile and read the datafile that way if
so.  This is the same fix we did in mapreduce/ to view the status
html.

Auditor: kamens,alpert,john

Test plan:
Rn a python and did
   sys.path.insert(0, 'genfiles/third_party.zip')
   import third_party.babel.localedata as l
   l.__file__            # confirmed it's in a zipfile
   l.load('en', False)   # works now!