Last place in babel we need to add in support for babel being in a zipfile.
Summary:
In production KA, all third_party modules are in a zipfile.  This
makes for special-case code when loading datafiles.

We don't use get_global() for anything yet -- and may never, it's only
used for timezones it looks like -- but best to fix it up so we're not
surprised later.

Test Plan: Fingers crossed.

Reviewers: ben, alpert, john

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D2158