extract_javascript is a built-in, so add it to the built-in list.
Summary:
I don't know why it wasn't there to begin with.

In theory this should't be necessary, but marcos was seeing
   ValueError: Unknown extraction method 'javascript'
This is probably an issue with pkg-resources somewhere, but that's all
black magic to me, and this seems a 'clearly' proper fix in any case.

Test Plan:
Ran
   % python
   >>> import babel.messages.extract as e
   >>> list(e.extract('javascript', open('/dev/null')))
   []
Before this change, it returned:
   ValueError: Unknown extraction method 'javascript'

Reviewers: john, marcos

Reviewed By: marcos

Differential Revision: http://phabricator.khanacademy.org/D3561