Update cldr localedata from 1.7.2 to 23.1 (!)
Summary:
pybabel 0.9 (what we use) can't parse such recent cldr files, so I had
to download pybabel 1.0 to do the parsing.  Note we aren't uploading
to pybabel 1.0 -- that's a lot of work, considering all the local
patches we have -- we're just using it to create a new localedata
directory.

So I did this in /tmp:
   % git clone git@github.com:mitsuhiko/babel
   % env PYTHONPATH=$HOME/khan/webapp/third_party/ python setup.py import_cldr

(The PYTHONPATH is to give access to pytz/).

I then replaced our localedata/ with the new localedata/*.  I ran the
following on both the new and old localedata:
   python -c 'import pickle; print pickle.load(open("babel/localedata/ak.dat"))'
and compared the output manually to confirm that every key in the old
localedata is in the new localedata as well (included for nested
dicts), and that the values are seemingly the same.

Test Plan:
I temporarily rolled back 9c8196ab378458d98dcb17b1d601faab1e486f65,
and I ran
   % python setup.py test

There were 27 failures, but they all seem ok to me.

1) unknown locale 'dv_MV' -- this has gone away, dunno why but don't
really care

2) Stuff like
Expected:
    u'Apr 1, 2007 11:30:00 AM'
Got:
    u'Apr 1, 2007, 11:30:00 AM'

3) We don't abbreviate as much anymore:
Expected:
    u'3 mths'
Got:
    u'3 months'

4) We lost timezone names:
self.assertEqual('11:30:00 AM EDT', string)
AssertionError: '11:30:00 AM EDT' != u'11:30:00 AM -0400'

(4) is due to using the pytz that's included in the babel distro; when
I forcibly force it to use the one in webapp/third_party/pytz, the
timezone tests pass.

Reviewers: james

Reviewed By: james

Maniphest Tasks: T1715

Differential Revision: http://phabricator.khanacademy.org/D4293