Undelete all the es-*, pt-*, tr-*, and da-* files, to get country code info.
We want to tell people from mexico to view ka in spanish, but to do
that we know that mexicans speak spanish.  The file that tells us that
is es_MX.dat, so we can't delete that file after all.  Extrapolating
this principle, every language we support a 'full' language experience
for, we want every lang+region data file that we can find.

For now, we only support 'full' for spanish. but pt is next, and da
and tk are lined up, so I just did them all now.

James points out: there was that xml file that we could use for the
country language map as well as another approach we could always fall
back to.  Other things we could do too.  But this is easiest. :-}

Auditors: james

Test Plan:
tools/runtests.py i18n_test.BestLanguageTest.test_babel_with_region