Removed es-US.dat
We are using the dat files to signal which countries spanish is spoken in.
We will probably have to go to another approach.  Until then, I'm just
removing the file for US, for while Spanish is spoken in the US, it is not
the main langauge, and we don't want to show the banner suggesting it.

Test Plan:
new test I am about to check in which makes sure there is no suggest language
for country code = US

Auditors: csilvers