Join python multiline comments by ' '. Each tag starts new comment.
Summary:
Just like we did for javascript in:
http://phabricator.khanacademy.org/rBAe053e691f92c32e3cf9f9bd82facf121c80d6f09

We now will merge multi-line comments onto a single line for messages
extracted from python too.  This is an issue for us since we sort the
comments in order to make the diffs cleaner, and unless the comment
lines were in alphabetical order they got scrambled up.

While at it I also fixed a spelling mistake in one of the test files.

Test Plan:
cd third-party/babel-khansrc
make standalone-test
It spits out 3 errors, but that is a known issue:
https://github.com/mitsuhiko/babel/issues/106

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D13207